perbaiki diskon

+++ resources/views/pelanggan/pelangganTambah.blade.php
@@ -447,7 +447,7 @@
 <script>
   function hitung(){
   var harga = parseInt((document.getElementById('harga').value).replace(/,/g, ''));
-  var diskon = parseInt((document.getElementById('diskon').value).replace(/,/g, ''));
+  var diskon = parseFloat((document.getElementById('diskon').value).replace(/,/g, ''));
   var dp = parseInt((document.getElementById('dp').value).replace(/,/g, ''));
 
   var totalDiskon = harga*diskon/100;
